remove all pre test stuff from core tests
